For the class definitions of RTC_DS1307; RTC_DS3231; RTC_PCF8523; and RTC_PF8563, added 'public' for the RTC_I2C base class. This gives the methods in 'protected' access in any inherited classes.
In RTC_DS3231.cpp, removed the check for INTCN bit in the control register, as setting an alarm is not dependant on the INTCN bit being set. The bits A1F and A2F are set when the alarm is triggered, regardless of the INTCN bit state. This allows setting alarms without needing to enable the interrupt control bit as the user may be using the SQW signal on that pin (e.g. 1 Hz).